ebook/requirements.txt
reportlab
Pillow
markdown2
python-frontmatter
PyPDF2
